plan - CRUD updates V0.5

+++ src/main/webapp/WEB-INF/views/plan-all.jsp
@@ -25,7 +25,7 @@
                 <td>${plan.name}</td>
                 <td><a href="/home/plan/details?id=${plan.id}" class="custom-button">Details</a></td>
                 <td><a href="/home/plan/update?id=${plan.id}" class="custom-button1">Update</a></td>
-                <td><a href="/home/plan/delete?id=${plan.id}" class="custom-button2">Delete</a></td>
+                <td><a href="/home/plan/delete/confirm?id=${plan.id}" class="custom-button2">Delete</a></td>
             </tr>
         </c:forEach>
         </tbody>
